Part 2 of Auth flow
This commit successfully completes the basic OAuth flow with Cloud
Foundry's UAA system.
The user navigates to #/login and where they are redirected to the
Cloud Foundry UAA page. The user inputs their credentials and will
login and redirect to the dashboard.

This commit also documents how to register the web service as a client
with UAA.

TODO: move to use environment variables
TODO: make a manifest that can be used to deploy the service.