Migrated OAuth Middleware redirection to the frontend.
Now the backend will only return {"status": "unauthorzied"} instead of
directly trying to redirect the users browser. Previously, the
redirection would try to occur on an angular $http call. This call does
not affect the browser. Instead now, the backend will return an
unauthorzied code and then the frontend will interpret that as a
necessity to redirect to the login page.