TEMP FIXME for sessions.
Previously, the service used a cookie store to save session information.
However, in production, the information saved in the cookie would
exceed the 4096 byte size limit supported by browsers. This commit
instead uses a FileSystemStore. This is a temporary fix that uses
FileSystemStore that stores the session on the local machine. This is
just a short term solution and should be migrated to a database.