Login view functionality
This creates a new component for login. It also sets the route
to check if logged in and go to login page(automatically now).
Eventually, it will check the actual status and go to the login page.

Another way to do this would be to use state in the react views rather
then using the router to render something. So the app view would get the
state for whether logged in and would render itself accordingly if or
if not. This solution right now seems simpler, but maybe as the
components grow that solution will make more sense.

Also used some basic tests for the view.