Fluxify the login functionality.
Create actions for fetching and receiving the login status. Eventually
a store would hold these.

I feel like making a login store is too much structure, in that all it
would be holding is the auth status anyway. It might make more sense
to have a more general store, but not sure what it would be.