Ensure store waits until authenticated
Had a little trouble with getting flux dispatcher working with
waiting for the auth status. Essentially, when the navbar calls
fetchAll(), the dispatcher is saying that auth status is still being
dispatched. It seems to be saying the STATUS_RECIEVED action is still
being dispatched, which I don't really understand because that should
already be done. I feel like I want to put a waits for in the fetchAll
action rather then in the Store code but this isn't possible. I don't
think this is the best way to solve this, so TODO ensure to look at this
again.