Added org memory limit api fetch
Same as org memory usage in how it will be used.

I'm differing from angular code in that I'm not using the url
defined on the orgs metadata, I'm hardcoding the url in the
method. I'm not sure if this is a problem. All the other ones
are hardcoded so not sure why this one wasn't.