Get service list partially working.
The service list will need a tailored table rendering as it requires
some properties to be nested two deep and has the third column which
is just a delete button. My plan is to render the whole table more
manually, a support operation with Reactable.

Also got teh tabnav working with new currentPage code.