Added docs folder and design brief doc
On other 18f projects, we set up a docs folder to put all md docs
in. This was done rather then the wiki to ensure we have version
control for these documents and to ensure devs have the docs while
working locally.

I've written a design brief to help decrease friction for new designers
who are coming onto the project as well as being upfront and clear about
what we're looking for. This is meant to make it easier for designers to
get started as well as ensure the designers know what their
responsabilities would be on the project.