App store
I changed the merge function so it's on the base store because the app
fetch method techincally requires it. Essentially for any store, when
you get an app, you should update it. The better, more "REST" way would
be go just replace the found old one with the new one. But this isn't
being done because of the weird way the API works in that getting all
doesn't give the same data as getting one.

The store is fetching again to correct provide waiting functionality.