Fetch based on current tab
This will fetch org or space users based on what the current tab is to
ensure we're fetching the correct data at all times.

One issue with this is that it does a fetch each time, which may not
be needed. This could be figured out later in the store.