Move user list to own component
Separates the user list from the page where you do user list navigation.

This was done to separate concerns so the list can act on it's own and
easily be modified separate from the more "view" type page that has
subnavigation for the user list.

This could be a pattern to have for all the lists.