Worked on getting tests back up
Theres still a failure on the new tests because the users requests are
taking a long time right now for some reason and the timeout isn't long
enough. Please advice on best way to deal with this.

I think I'm at the point where we should merge this branch. A lot of
the testing things I'm fixing now are more global then just this story
so it probably needs its own story.