Service plan store and tests
I made the service plan a separate store to make it simple to use
and build.

I made a separate method to filter service plans by service because
the UI will be to go back and forth quickly between different services
to see their plans. At some point, the data might be merged somehow
so the same service plan can belong to multiple services and this will
make that easier.