Fixed CF API to not hide errors
I figured out what was hiding errors, using `then` in the API code
for promises rather then done. I'm using a polyfill for `done` that
will ensure errors come up. I can confirm this works because I threw
an error in one of the API methods and it correctly showed up in
console.

Also finished marketplace, am nesting a second table under the first
as the simplest way to make the UI right now, as I'm waiting for
design feedback still. This required making a different table that's
not the Reactable, as Reactable didn't support nested tables.

The main route function should request services straight from the UI
rather then use the action to stay consistent and be more simple.