Modifying roles on the server.
I changed the helpers setupViewSpy to a stub. This was to minimize
the impact of the tests. So it will only test that the method was called
and the call stack wouldn't go further into the store code, which is
tested separately. Before this change, if a action listener in the store
failed on a action test, that action test would fail which doesn't make
sense because only the store test should fail in that case.