Finished UI with handlers and some mapping
The main 3 roles used need to be mapped from being called on the backend
to what they appear as in the front end models. So when you want to
delete the auditor role from a user, you have to use "auditors" when
making the API request. But when you get the user model from the server,
this same role is called "org_auditor". I made mapping in user store and
user list control to take care of this confusion. The mapping just
switches the data for the correct context.

This obviously isn't very future friendly. The current way of just
getting the all users again whenever a role changes isn't great for
performance either.