CSS build
This CSS build uses CSS modules to scope CSS classes to the current
css file that gets imported. This reduces "global" css and could
make it easier to maintain, bundle and share each component.

I'm currently configuring all the bootstrap or 3rd party css code
to be not compiled with css modules. This is so I don't have to go
through and change all the bootstrap classes to be imported.

In the future, when bootstrap is removed, I may set it up to have
all code, 3rd party and local, to use this system and have no
global css.