Upgrade to babel 6
Babel 6 improves:
- performance of compilation and code
- modularization of compilation components, different transforms are
in plugins now, allowing a plug and play approach. This could allow
further optmiziation if we want to be very specific with what
features from es2015 we want
- ensures we're up to date with js standards updates.

Currenty there's a bug:
https://phabricator.babeljs.io/T6644

which is prohibting the use of the babel runtime transform. This
means there will be some extra code when compiling with webpack.
Once that bug is fixed (a lot of people are waiting), then chang-
ing this line will bring back runtime to lessen compiled file size

```
plugins: ['transform-runtime']
```

Still TODO:
- bring the code to pure es2015, get rid of stage-0 code.