Add eventually attempt to fix timing issue
The checkbox takes some time to switch states due to an ajax req
that happens, so this is an attempt to ensure it's waited for
before testing.

Don't actually think this will work because it's finding multiple
checkboxes but wanted to see the test run on travis, where it's
breaking.