Working with all images and fonts
Now all images and fonts in both cg-style and uswds are using an
variable for the path, which I'm modifying in the data section of
my import. TODO: should this be moved to cg-style somehow?

I also had to reconfigure the url-loader so webpack knowns where
to find it, because when the base.scss file is being processed, it
won't know where that loader is.

The limit here is being set so only files over it will be set as files,
anything under will be inlined in base64.