Updated code workflow
I think it's technically fine for anybody to merge once it's ready for merge. This ensures if the reviewer is blocked by the PR, they can merge it after they review. If they'd rather not merge as they're not blocked and want the original writer to deal with merging and the deployment, they can hold off on merging.