Use webpack to load css modules from cg-style
To load CSS modules from webpack, it needs to be configured with the
proper loaders and the proper paths. Because cg-style is still under
active development, we don’t install it from the npm registry.

npm link (by itself) isn’t sufficient as its just a symlink. When
webpack goes to try and resolve things, it will use the absolute path
so we need to include that in the loader.

That path has been abstracted away into an environment variable called
CG_STYLE_PATH, which should be an absolute path to the location of the
cg-style repo.

The README has been updated to also reflect the new environment
variable.