Refactor navbar component to use cg-style sidenav
Now that cloudgov-style is being pulled into the project and webpack is
handling CSS modules, it should be used on the navbar component. The
render function was refactored to account for the new HTML structured
used by cloudgov-style.

Use the orgActions to toggle menus’ display state so that we can keep
the component as stateless as possible and let the store be the one
source of state.