Update guid state of space component on changes
The space component relies on a “currentSpaceGuid” to ensure that it
displays the correct information to the end user. The change handler
was not correctly setting the state of the component which resulted in
the inability to switch from space to space without refreshing.

Also make sure to cleanup listeners when components are removed.