Add CG header related JSX components
Added new disclaimer and header components to show the USA flag
disclaimer and the typical cloudgov-style header. I added them as two
separate components because React components need to return everything
encapsulated in a single root node but the disclaimer HTML and header
HTML shouldn’t be in the same root object beyond the app root.