Add datetime formatting util function
Provide a utility function to format date strings in a consistent
manner across the application. The function takes a timezone (that
defaults to UTC) so that we can potentially allow users to toggle
between timezones in the future.

Unit tests are added to ensure that the function guards against bad
input and to ensure that the returned string is properly formatted.

Also includes the addition of `moment-timezone` to `package-.json` and
adjustments to `webpack.config.js` so that the timezone data is loaded
properly.