Use formatDateTime() in <ServiceInstanceList/>
Use the formatting util function in the “Updated at” column in the
service instance table. Since we don’t currently have a mechanism for
the user to supply a timezone, the timestamps are formatted using `UTC`
as the timezone.

If the service has not been updated, then the `created_at` timestamp is
used instead.

Linting complaints were fixed.