Added new stat data to app store
I can't think of what to do if the stats are there for an app
that isn't there yet, because the data would be uncomplete, so
just throwing an error.

Just though of a problem: what if the app stats request gets in
before the app summary request? Should probably change how the
requests are done.