Change BaseStore to use immutable data structures
The underlying data structure of BaseStore is now an immutable list,
which is similar to a regular JS array. An important thing to note is
that any `get…()` method converts the immutable data back to regular JS
objects. This means that stores use immutability internally but that
anything that retrieves data from a store doesn’t need to care since
they are converted back to JS objects.

I also changed the API of the BaseStore a bit to make dealing with
immutable data a bit easier.

I added a `.push()` method which takes an argument, converts it to the
respective immutable data type and then creates a new list with that
item pushed on and sets that array to be `_data` on the store. It also
emits a change event as the underlying data has changed.

One of the benefits of using immutability is that we can easily compare
nested objects for sameness. This means that we can quickly/cheaply do
checks to see if data has changed or not and act accordingly. I added
`dataHasChanged()` to the BaseStore which returns a boolean. This gives
every store the ability to only emit change events when the data has
actually changed.

I also added two new merge functions that account for the fact that the
data is immutable. A nice benefit is that these merges no longer have
unexpected side effects.

There are two new merge APIs that have very similar function
signatures. Each takes three arguments, with the last being an optional
callback. The first argument is the key by which to identify similar
objects and merge them (generally speaking this is `guid` in our
project), the data to merge in, and a callback that is passed `true` or
`false` depending on if there was any data that changed. `merge` takes
a single object as the data to merge and `mergeMany` takes an array of
objects. By default, the callback will call `this.emitChange()` if
there is a data change.