Add .delete(guid) method to BaseStore
There is a common need across the store objects to be able to delete an
item which matches a `guid`. I refactored the code into BaseStore so
that it is available for all stores.

I am using the callback pattern that we have in `merge` so that there
is a consistent API for stores. I return the callback in all cases, and
the default callback returns nothing. It just emits a CHANGE event.
Thus, unless a callback is supplied which returns something the
returned value of `.delete()` isn’t going to be very valuable.