New route actions
New actions to get all routes for an app. These were separate methods
so that the appGuid could be passed along. The route data that comes
from this request doesnt' include the appGuid that the routes belong
to, so we pass this guid around so the route store will add it to the
data structure. This method of doing this is consistent with the rest
of the codebase, and ensures only the stores change data, rather then
the API doing it. It also means consistency, as if routes are gotten
in a manner where there will be no app guid, then the store has to
handle that case.