Route list view
Uses an normal table as I don't really need the functionality of
reactable and I want to get rid of that library anyway.

I do the filtering of the routes so I only get the routes thet bleong
to the current app in the state setter in the route list component. This
is a little different but might be a good solution rather then adding
more methods onto the store.