Include login/logout link in site global nav
Quick sketch of what using a header link for login/logout
might look like in the global nav. I substituted the
previous "Deck" link with a link at the far right of the
navigation because user testing revealed that the "Deck"
link was of very little value to users.

The link will be "Login" when the user is not yet
authenticated and "Logout" when the user is properly
authenticated.