Route store handling domain
The route store's guid has to be matched with an existing route
domain_guid. If no route with matching domain_guid exists, it should
not do anything.

Rather the add new functionality onto base store, this was done by
constructing a specific object to use the merge function with. This
object is created with a `domain` key that holds the whole domain
object and then a domain id which is just the domain's guid.

The test seems to show that if the matching route doesn't exist then
no new routes are created, which is correct.