Merge dependencies and devDependencies
To close #285, I have merged the two sections of the `package.json`
file and regenerated the `npm-shrinkwrap.json`.

I did this because it is important that the deck development team
always know what version of every dependency is being used in the app
AND for the deck development team to have a simpler understanding of
dependencies.