Updates to getting started instructions
The set up instructions were a bit out of date for the
`staging-alpha`/react rewrite and didn’t include all the steps for
getting started.

Mostly, this added some instructions on cloning the repository so that
it fits with a typical Go project structure. I also changed references
from Angular to React so that all the information is accurate for the
rewrite branch.