Scroll to service instance creation form
When the “Create service instance” button is clicked, the form is added
at the bottom of the page, which gives the impression that nothing
changed if you are near the top.

I added a `.scrollIntoView()` method that will scroll the form
component into the user’s viewport. This method is called on any change
to the underlying store and when the component is first inserted into
the DOM.