Add delete confirmation to instance
Keeps state for a service instance for when teh confirmation UI
is there as a simple boolean on a `confirmDelete` key. This allows
the UI to check whether a confirmation box UI should be shown for a
particular instance. The key will be deleted on the cancel event
in the same way. This keeps the UI state with the store data.