Init components with existing data.
This changes components inits to ensure they all start with existing
data in the stores.

In one case, the marketplace, this fixed an actual bug where the
data wouldnt' appear at all. In other cases, it makes things speedier
on the second time you look at the data on a page.

This still sends off requests to get the data, meaning that if it's
changed, there will be a re-render when the ajax request comes in.