Change current org for marketplace view
The marketplace view wasn’t ensuring that the current org was properly
set, and therefore if you loaded the marketplace URL as the first view
the value won’t be set properly.

This breaks the service instance creation form because it filters for
spaces that belong to the current org.