Show space name in <Space />
It was a bit confusing to see just “Space” so we wanted to have the
space name displayed as well. It turns out that the code was there, it
was just buggy. I changed it to use the `name` key on the `space`
object in state instead of looking for a `spaceName` key (which doesn’t
exist).

Refs #384