OrgStore emits change when the current org changes
If the OrgStore gets a ORG_CHANGE_CURRENT action it changes the
currentOrgGuid. But it wasn’t always calling emitChange to update its
subscribers. This was presumably to protect against the orgGuid being
changed to one that didn’t exist in the store. However, this meant that
if the store had not yet been fetched, the change event would not fire.

I changed the handler so that it always changes the current guid and
calls `.emitChange()`.