Problematic implementation of loading states
Right now, it is hard to use a single `fetching` value on the store
because a store might have a few different requests that have to be
received before the full state can be constructed. For example, on the
app page we need to not only get the app but also the app states and
routes. If we only have one attribute on the store, then it is hard to
time when the attribute gets set to false or true