Fix tabnav accessibility.
Uses http://heydonworks.com/practical_aria_examples/#tab-interface to
create an accessible react tabnav.

One additional item is to add the `aria-selected` attribute to the
tab nav items. This is harder to do because they're already created
by the time it gets to the tabnav class, but might try and do that
later.