Use new APP_ALL_* events in store
We have two new action types to help us coordinate when we have
multiple requests in flight. We want to know when all the requests have
been made and when the responses have come back so we can set the
`.fetching` attribute on the store appropriately. This will be used for
showing loading states in the UI.