Remove privileged proxy for converting ids between e-mail addresses
Originally, you needed the user id to set roles. Back when this was
corrected the only way to do that was to query for the user info first
by e-mail address to get the id. Now, there are APIs that make it
possible to set roles by e-mail address. This patch removes the need
for the privileged query call