Stores and actions for changing user viewed type
Currently, the current user tab (org or space) is being tracked directly
in the users.jsx component, but it can be moved to the store as all
state should be tracked in the store. This will make it easier to
keep track of what's going on and when things should change.