User store keep tracking of current user
The current user guid will be kept in an attribute if it's found
and the get method will get it with from the existing users in the
store.

The likely case is the user won't be in the store yet, how should we
fetch the user and get the correct permissions?