Fix authorities and add autoapprove
Addresses https://github.com/18F/cg-deck/issues/340

Prior to this patch, an admin would setup the cg-deck client with authorities which means the deck would have it's own permissions. We don't want that. Instead, we want the client to only use the perms the user logged in has.

Additionally, the current set of scope is a subset of the scopes used for the cf CLI.