Fix logic mistake
Changes logic so space menu will open always. This means that you can't
close an org menu in the list. I think this is better for now then
having the menu open and close rapidly while rendering.

In order to fix this logic, we'd have to look at ensuring the main
routes don't toggle the menu action but somehow initialize the orgstore
with the correct org selected.