Use single css file for all components
CSS modules is a cool idea, but our implementation got a bit
complex. We were importing the individual component files
from the style library, meaning that anybody working on a
component had to know not only that a class was in cloudgov-
style but also which file it was in. This was a lot of work
for something as simple as adding a class.

Now, all components just use the final, generated output css
from cloudgov-style so that any class in the library should
work pretty well in any component. I kept the .styler()
convention mostly in place to minimize changes.

The best illustration of how this changes improves code
legibility is in navbar.jsx right after render() begins.
Where we used to have to know which component file a class
was in, now we can simply use any class from the style
library.

Refs #350