Track Google Analytics only on production
We don't want to pollute our user data with
development data, so Google Analytics
should only be used in production.

Check for the environment variable
NODE_ENV, and if it is present, then track
events in Google Analytics. Otherwise, we
do not.

Idea from a Dan Abramov tweet
(@dan_abramov/status/592692202335301636)