Add mergeAll to base store
The merge algorithm differs from the one in .merge() as this will merge
the new data into all entities in the store that match on the merge key
and not just the first one.

This is useful if you want to update many entities in the store based
on a merge key and some new data, such as when we get a domain that
might apply to many routes.