Add dashboard uaa icon to public URL
In order for cg-uaa to have the correct dashboard icon on it's pick
a place to go page, we can add it to an accessible URL for the
dashboard.

These sticks with the webpack build process so nothing there has
to change, but includes an awkward import in `main.js` to import the
file even though it's not used in this project.

Ref 18F/cg-uaa#25