Set fetching after data updates in stores
There was a weird space between the time when fetching would be set to
false and when all of the data had been merged in. Since both emit a
change event, the components would re-render and could result in
flashes of the null state (no apps, no services, etc).

By setting false after the data update is complete within the stores,
and only emitting one change event (which is triggered by the change in
the value of .fetching), we eliminate a potential reason for flash of
content bugs.