Fix user store so tests pass
We had previously been setting .fetching at the top of each
of the handler functions in the stores. The problem with this is
that changing the fetching value will trigger a change event,
which might fire before the data that was retrieved from the API
has been merged into the stores. The result is a flash of content
like bug where the user sees a flash of "No apps", etc.

It looks pretty weird to set the value in two places like this,
but I set it after the merge data operation has been completed
and in the case that the request retrieved no data.