Update route store on APP_RECEIVED event
When an APP_RECEIVED event is fired, the app object that comes along
with it contains all the route and domain information we need. So on
that event, we can update the route store instead of only relying on
sending a subsequent network request. If the app object doesn’t contain
a child routes object then we will just ignore the event.

Also changed the appGuid key in the route store to app_guid to be
consistent with the kinds of data structures we keep in our stores.