Added fetched property on stores
Fetched can be used to determine if a store is empty because the
data wasn't fetched yet or because it was fetched but there is no
data. This distinction is needed for the components.

Implemented the same as fetching because I think consistency is more
important less code. A question to ask here: is conversion to boolean
important enought to justify get/set methods? I think either way that
should be refactored later.