Implement fetched in app store
Have the app store use new fetched property by using it to check whether
the app should be shown as currently empty.

I created a separate function in App component as there's two places
that have to check whether an app has enough data; the check for the
app being empty and the app being ready to render.