Build with Javascript sourcemaps in development
To make debugging easier, I enabled javascript sourcemaps in
development. The mechanism relies on the NODE_ENV environment variable.

If the environment variable is set to “prod” then webpack does some
optimization on the bundle file such as minifying it.

If it is not production, then webpack should also generate sourcemaps
which make it much easier to track down bugs in the compiled/transpiled
code.