Refactor routes to separate space pages
Refactors the routes on the space page to be separate functions for
each page on the space page: apps, users, services. This allows to
to specific thing for each page, such as specifically fetching things
on each. It also allows to have another potential page on the users page
for org vs space users.