Configure karma to only report test failures
There used to be a lot of things that would get printed to
the console when the tests were run. It was capturing
console.log output from the console and showing tests that
passed. All of this was a lot of information, and made
finding failing tests a bit more cumbersome.

I am now capturing all console.log output so nothing gets
reported and have introduced a new karma test reporter
to only show when a test fails.