Use timeout on loading and reconfigure stores
Uses a timeout on loading so the loader only shows up after a req
has taken more then 400ms. TODO: could this logic be in the Loading
component so it's shared.

Changed back marketplace so it listens to both sevice and plans store
for them having loaded completely. Reconfigured the service plan store
so it keeps track when it makes the array of requests for plans. Now
it will only say it's been fetched when all of those requests go
through. This method also allows the normal service plans fetch route
to work by using a separate property to track multi request.