Fix short break in loading
Before there was a short break in the loading component between the
services resv and service plans fetch actions. This was due to service
plans not having set fetched before services unsets it because they both
listen to the same event and services came first. Easy fix was to add
an unused request to fetch service plans to set it's fetching to true.

This isn't great because it doesn't make logical sense for that store
path to emit a change if it's not making the request.

Other options
- Have hte marketplace component directly set fetcheing on service plan
store. Kinda cheating the flux way though.
- Somehow use waitsfor on service store to wait for plan store.