Refactor button component
Refactor the button component so it's a generic action component that
can be a button or a simple link.

This component can be configured to
- be disabled,
- be a certain type of style,
- have extra classes such as testing hooks,
- be a link, button or submit,
- have a label

If there are style changes for the new buttons on the new panels,
we shoudl update the style for those globally, so everything uses the
same styling.

Question, for the styles, should we be defining them so precise to
the actual styles in cg-style? For instance, rather then making a button
of stle "outline", maybe we should make a button of style or function
cancel, or confirm. So the vars are more based on functionality then
their styleing?