Fetch all spaces on initial load
It turns out that the endpoint `/spaces/:spaceGuid/summary` returns a
different set of information than `/spaces/:spaceGuid`.

The `/spaces` endpoint returns a collection of information that can be
found at the non-summary endpoint. This data is important because it
determines if a space has a quota attached to it.

Adds constants, actions, API calls and store handlers to fetch all
spaces, process them, and merge the data into the space store.