Add domain store
Domains are a distinct entity from routes, even though they are
linked. Before, we were just fetching domains and storing them
on matching routes. But now that we need all of the possible
domains available to us so that we can enable route creation
it made sense to isolate domain information in its own store