Add new route actions, constants and store handlers
To enable users to create, read, update, and delete their routes we
needed a whole slew of new constants, and thus actions and store
handlers for those actions.

Also did a bit of organizing, so a lot of this thrash is due to me
alphabetizing things within files for my sanity.