Fetch shared and private domains separately
The main domain action defaults to "private" as this isn't used yet.
The route store will check if the route uses a shared domain route or
a private domain route in the domain url and then uses the appropriate
cf api method. This fixes the bug where shared domains would never be
fetched, such as `apps.cloud.gov` domains.