Route creation fails when no path
Because the code is assuming theres a value and it's an array with
something in it, but in reality, if you don't set the path, then
the path is undefined, which is now checked for in a if statement.

I created this as a separate branch of the if statement because the
current one was getting log and confusing and would default to the
path having a `/` which is invalid in cf if you set the path to just
a `/`.