Move routes list to routes panel
To keep consistent with how services is working and because the UI
is changing so it's not within the app settings panel anymore.

I decided to not pass any props data to the routes panel and keep all
data in the stores on that panel. This was for consistency, as it's
easier either make it a props component or a store/data component.