Fix route form from submitting to change url
There was a bug where clicking cancel would submit the form, event
though we were capturing the event and preventCancel.

The issue was we were setting the button type to "outline" THis was
due to the old way we were using the "type" property on the <Action>
component. We since changed that prop to be based on style, it should
not be tied at all to the HTML button "type" attribute.