Dont format summary api responses
CF api is inconsistent in that responses from "summary" api calls
are not split into metadata and entity but just have all the data
on teh main object. So don't format the reponse for them.

Found the best way to do this was a simple URL check. Another
possibility is to do various checking of the object before formatting,
such as if it has entity, metadata, if it has the guid on first level
of object. That solution seems more robust but also complicated, which
is why I opted in for the simpler one here.