Use separate test node env var
Separated from prod env var which we were using for all circle tasks.
The problem with this is it requires a separate build to happen
on production. Another option woudl be to conditionally set an env
var in the "dependencies" seciton, but I can't find anything about
this being possible in circle.