Count status componenents
The count status components are for showing a count of total
and status of all apps, spaces or potentially any other thing that
has status.

I created specific components for apps and spaces and then one generic
one that they both use because there's enough differences between
the two that it's easier to contain the logic in separate components
this way.