Pass worst app status to status count.
In order to give the dev an idea of problem apps when quickly looking
at app count, we'll pass the worst status, so they'll see crashed,
then stopped, then running in some form.

I used a method that gets the highest ranked state on the app count
status. Can also move this to util if it's needed again.

I had to add a method to main to get the full space for the first
10 spaces. We decided on around 10 for performance reasons, as more
could be a huge request on the first page. The space fetch here gets
the full app objects for a space and place them in an "apps" array
on the space, which is used in the org quick look to pass to the
app count. The cf api method for spaces fetched was changed so that
it returns the actual data so it can be used in `then()` methods.