New action for toggling org quicklook.
This is a new action separate from toggleSpaceMenu because it has
to be kept as separate in the data, otherwise the menu would also
open up when you open up the quicklook.

It'd be nice to find a way to consolidate some of this code in the
future, but this seems appropriate right now.