Orgstore change when toggling quicklook.
The org store will toggle a new proper, quicklook_open when the
quicklook action. This is separate from other org "open" toggles,
such as spacebar, so they can both be opened independantly.

The quicklook toggle is a true toggle, unlike the space menu, due
to the UI requiring multiple orgs allowed to be opened at once.