Basic space quicklook layout
The space quicklook uses nested panel rows to show each space and
each app for each space. The nested panel rows made most sense as
the UI for each has the same requirements. The nested rows will be
styled based on where they are in the hierarchy of nested panel-rows.