Add color coded app status
If an app is crashed, shows the app name in red and the state. Otherwise
just color code the app state.

I didn't use an extra component for this status text as I actually
couldn't find it being used anywhere else. It could potentially become
a status if theres another use case for it.

There's a new class for this, which pretty much just changes text color
and ensures the state isn't uppercase as it is in the API.