Correctly show loading state for space quicklook
The space quicklook should show a loading sign if the spaces are
still being fetched or loading. This is tracked in one place at
the overview page level and passed through as a prop. Additionally
there has to be code so the whole overview page itself doesn't load
when one org is fetching all it's spaces.