Org quick look icons
Adds icons to the org name, and space and count status components.
A new component for the entity icons was created to reduce certain
code like setting the right status class, ensuring a dev doesn't
pass the wrong icon type (can only be space, app, org), and setting
defaults for icon color. Refactored the Icon component so it uses
styler interface to keep consistent with newer code.

There was also functionality added to differentiate between stroke
and fill icons. This is a limitation in the SVGs, some of our icons
really have to be stroke and some really have to be fills. The current
set of icons are fills while the previous ones are strokes. I'm opting
to keep the generic `icon` class as stroke icons and adding a new class,
`icon-fill` for these new fill icons so that this isn't a breaking
change for cg-site. I think this is a sane default to set anyway.
Unfortunetely, a dashboard developer will have to know whether their
icon is a stroke or fill when using the Icon class, which is not a great
interface. Lucikly for entity icons, EntityIcon takes care of it.

Requires 18F/cg-style#ms-entity_icon branch on cg-style.